requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.0
Pillow>=10.0.0
pillow-heif>=0.13.0
pyperclip>=1.8.0
